app/requirements.txt
httpx
gradio
weaviate-client>=4.0
pydantic-ai
